[Feat]: 유저 포인트 충전 차감과 내역 저장

+++ src/main/java/io/dev/concertreservationsystem/application/user/UserAdminDTOResult.java
@@ -23,7 +23,7 @@ public record UserAdminDTOResult(
     UserGenderType gender,
 
     @Min(0)
-    Integer point,
+    Long point,
 
     @NotNull
     LocalDateTime createdAt,
