+++ src/main/java/io/dev/concertreservationsystem/domain/user/UserDTOResult.java
@@ -33,7 +33,7 @@ public record UserDTOResult(
         @NotNull
         @NotBlank
         @Min(0)
-        Integer point,
+        Long point,
 
         @NotNull
         @NotBlank
